+++ history.md
@@ -1,3 +1,4 @@
+* Added Stream music mode
 * Fixed bug when lancemate with pet quits party
 * Fixed bug when reading black market sign with a pet in your party